+++ src/controller/register-controller.js
@@ -46,13 +46,3 @@ export const validateBtnSignUp = () => {
     });
   }
 };
-
-export const validateBtnLogIn = () => {
-  const btnLogIn = document.querySelector('#btnLogIn');
-  if (btnLogIn) {
-    btnLogIn.addEventListener(('click'), () => {
-      console.log('cambiando el hash');
-      window.location.hash = '#/home';
-    });
-  }
-};
