+++ src/css/styleMain.css
@@ -25,6 +25,7 @@
 figure img {
     width: 100%;
     height: 100%;
+    border-radius: 10px 10px 0 0;
 }
 .flex-comun {
     background-color: tomato;
@@ -484,7 +485,7 @@ body::-webkit-scrollbar-track:active {
     figure img {
         width: 100%;
         height: 100%;
-        border-radius: 10px 10px 0 0;
+        
     }
 
     .new-post {
